made function def class
